Use ts-node to import modues.ts
